Daniel 10:21 capitalize book title
Daniel 10:21  "The Book of Truth" is a document title and should be capitalized because there is no better way (yet) to graphically represent titles. Opinion is about equally divided on whether this is a distinct title or the attribute of some writing, "written in the true book." But this rendering seems easier for MTTs to notice the meaning.