Esther 2:23 capitalize book title
Esther 2:23  "The Book of the Chronicles" is a document title and should be capitalized because there is no better way (yet) to graphically represent titles.
Also, I de-capitalized "the" before "Chronicles" and the word "presence" before "of the king" since both have no significance to require capitalization. Also "king" because this is a function and not a title tied to a specific name.