Nehemiah 8:8 capitalize book title
Nehemiah 8:8  "The Law of God" is a document title and should be capitalized because there is no better way (yet) to graphically represent titles. This phrasing, taken almost word-for-word from the ASV, is confusing without the designation of the specific document from which they read. So I also dropped the "in" before the "The Law of God."