Acts 11:5 & 6 simplify grammar and add period
Acts 11:5 Original: four corners; it descended to me
Change: four corners. It descended to me.
I'm following the suggested rewording by Suze Lau, one of the readers for the text team.
I also changed the semi-colon to a period in verse 5 and then also in verse 6 and made these clauses independent sentences to simplify the grammar.