Acts 10:23 eliminated extra words not in Greek
Acts 10:23 Original:  So Peter invited them to come in and stay with him and the people with him. On the next morning he got up and went with them,
Changed to: So Peter invited them to come in and stay with him (delete other words). On the next morning he got up and went with them,
I'm following the suggested rewording by Suze Lau, one of the readers for the text team.