22:17
Trance changed to vision. A trance is a state of altered consciousness (of doubtful authenticity). He was given a vision, so that he could "see" Jesus. This is a simpler term. The Gk is ἐκστάσει,  is means to "be taken from one place to another" with the figurative meanings. Vision seems less problematic and fits the rest of the verse.