Acts 11:5
Trance seems to be misleading. Trance seems to be an altered state of consciousness (raising the question of how authentic would a "trance" be for communicating an event. But being given a vision is closer to the meaning. ἐκστάσει is to take to another place, as it were. These are figures, and the nature of the vision does not belie unconsciousness or semi-consciousness. The vision seems to fit what is happening much better.