Acts 10:10
Changed "fell into a trance" this doesn't really capture ἔκστασις. The idea of a trance seems foreign to this "taken to another place" of ἔκστασις. This is more consistent with the description of vision. A trance is transcendental and sub-conscious or partially-conscious. The MTT would be served by using "vision" and not trance, in my view. Note most modern versions follow this rendering.