ωστε introduces consequence, not basis
As a result indicates that what follows in verses 13–14 is the
consequence of the progress of the gospel (cf. Bruce “the upshot of it
all has been this”). Even though the phrase has been discarded by many
translators, it is best to have it for a better connection.

Loh, I.-J., & Nida, E. A. (1995). A handbook on Paul’s letter to the
Philippians (p. 20). New York: United Bible Societies.

Also, I’ve made it plain that there are two results by combining the
sentences.