Update 03.usfm
v. 6 "become proud and fall into condemnation"
changed to:
"swell with pride and might into condemnation"
The UBS notes indicates that "become proud" is a metaphor referring to "swelling with pride" like smoke rising from a fire puffs up into the air. And, the verb for "fall into" is subjunctive, so "might" is added to bring out the action as potential.