Ezk 3:20
No version I consulted, nor NICOT, makes "I set a stumbling bloc" the
result of the preceding clause. All make it part of the protasis, and
only "he will die" as the apodosis. I have followed the placement of
the zaqeps and 'athnach for the logic of the second sentence.