Ezra 1:3 change wording
Ezra 1:3 change wording FROM: "Any one of you who comes from his people, may his God be with you, for you to go up to Jerusalem" to what you see because this is a very poorly constructed sentence. The Hebrew and the English translations I consulted used the 3rd person, not the 2nd.